requirements.txt
torch==2.3.1
torchvision==0.18.1
pillow==10.4.0
tensorboard==2.17.0
matplotlib==3.9.1
streamlit==1.36.0